requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
django-filter
Pillow